requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
httpx
beautifulsoup4
requests
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
dnspython
pytz
